+++ src/collective/dms/mailcontent/vocabularies.py
@@ -13,7 +13,7 @@ from zope.schema.vocabulary import SimpleVocabulary
 
 
 class EmailAttachmentsVocabulary(object):
-    """ Vocabulary listing outgoing mail files, as possible email attachments """
+    """ Vocabulary listing outgoing mail files and related ones, as possible email attachments """
     implements(IVocabularyFactory)
 
     def __call__(self, context):
@@ -22,18 +22,23 @@ class EmailAttachmentsVocabulary(object):
         brains = find(context=context, unrestricted=True, object_provides=[IDmsAppendixFile.__identifier__,
                                                                            IDmsFile.__identifier__])
         for brain in brains:
-            terms.append(SimpleTerm(brain.UID, brain.UID, _tr(u'({}) => ${{title}}'.format(brain.portal_type),
-                                                              mapping={'title': safe_unicode(brain.Title)})))
+            obj = brain._unrestrictedGetObject()
+            terms.append(SimpleTerm(brain.UID, brain.UID,
+                                    _tr(u'({}) => ${{title}}'.format(brain.portal_type),
+                                        mapping={'title': u'{}  (« {} »)'.format(safe_unicode(brain.Title),
+                                                                                 obj.file.filename)})))
         # then we find files of related mails
         pc = api.portal.get_tool('portal_catalog')
-        for rv in context.get('reply_to', []) or []:
+        for rv in getattr(context, 'reply_to', []) or []:
             brains = pc.unrestrictedSearchResults(path=rv.to_path, object_provides=[IDmsAppendixFile.__identifier__,
                                                                                     IDmsFile.__identifier__])
             for brain in brains:
+                obj = brain._unrestrictedGetObject()
                 terms.append(SimpleTerm(brain.UID, brain.UID,
                                         _tr(u'${{ref}} ({}) => ${{title}}'.format(brain.portal_type),
                                             mapping={'ref': rv.to_object.internal_reference_no,
-                                                     'title': safe_unicode(brain.Title)})))
+                                                     'title': u'{}  (« {} »)'.format(safe_unicode(brain.Title),
+                                                                                     obj.file.filename)})))
 
         terms.sort(key=lambda trm: trm.title.lower())
         return SimpleVocabulary(terms)